requirements.txt
ckanext-files==0.0.3
ckanext-admin-panel==0.0.1
